Extract TokioCompat adapter to shared harnx-acp::compat module
TokioCompat<T> — the bridge between tokio's AsyncRead/AsyncWrite and
futures_io's AsyncRead/AsyncWrite — was duplicated in three places:

- crates/harnx-acp/src/client.rs
- crates/harnx-acp-server/src/server_main.rs
- crates/harnx-acp-server/src/lib.rs (test module)

Extract it to a single canonical location: crates/harnx-acp/src/compat.rs,
exposed as #[doc(hidden)] pub mod compat so harnx-acp-server can import it
without the module appearing in public API docs.

Remove all now-unused imports (Pin, Poll, TaskContext, TokioAsyncRead,
TokioAsyncWrite, ReadBuf) from the consumer files.

#72

Plan: issue-72-tokiocompat-dedup
