+++ crates/harnx-acp-server/src/server_main.rs
@@ -5,14 +5,12 @@
 //! to `agent_client_protocol::AgentSideConnection`, and drives the I/O loop
 //! until stdin closes.
 
-use std::pin::Pin;
 use std::rc::Rc;
-use std::task::{Context as TaskContext, Poll};
 
 use agent_client_protocol as acp;
 use anyhow::{anyhow, Context, Result};
+use harnx_acp::compat::TokioCompat;
 use harnx_runtime::config::GlobalConfig;
-use tokio::io::{AsyncRead as TokioAsyncRead, AsyncWrite as TokioAsyncWrite, ReadBuf};
 
 use crate::HarnxAgent;
 
@@ -92,49 +90,4 @@ async fn run_local(config: GlobalConfig, agent_name: String) -> Result<()> {
 
     result.map_err(|err| anyhow!("ACP server I/O error: {err}"))?;
     Ok(())
-}
-
-/// Adapter between tokio's AsyncRead/AsyncWrite and futures_io's AsyncRead/AsyncWrite.
-/// `agent-client-protocol` takes futures_io traits; stdin/stdout give us tokio traits.
-struct TokioCompat<T> {
-    inner: T,
-}
-
-impl<T> TokioCompat<T> {
-    fn new(inner: T) -> Self {
-        Self { inner }
-    }
-}
-
-impl<T: TokioAsyncRead + Unpin> futures_util::io::AsyncRead for TokioCompat<T> {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut TaskContext<'_>,
-        buf: &mut [u8],
-    ) -> Poll<std::io::Result<usize>> {
-        let mut read_buf = ReadBuf::new(buf);
-        match Pin::new(&mut self.inner).poll_read(cx, &mut read_buf) {
-            Poll::Ready(Ok(())) => Poll::Ready(Ok(read_buf.filled().len())),
-            Poll::Ready(Err(err)) => Poll::Ready(Err(err)),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-impl<T: TokioAsyncWrite + Unpin> futures_util::io::AsyncWrite for TokioCompat<T> {
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut TaskContext<'_>,
-        buf: &[u8],
-    ) -> Poll<std::io::Result<usize>> {
-        Pin::new(&mut self.inner).poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut TaskContext<'_>) -> Poll<std::io::Result<()>> {
-        Pin::new(&mut self.inner).poll_flush(cx)
-    }
-
-    fn poll_close(mut self: Pin<&mut Self>, cx: &mut TaskContext<'_>) -> Poll<std::io::Result<()>> {
-        Pin::new(&mut self.inner).poll_shutdown(cx)
-    }
 }
